b24_tgbot/requirements.txt
aiogram==3.4.1
httpx==0.27.0
python-dotenv==1.0.1
pytz==2024.1